Nowa wersja testowa 1.99
